Add files via upload
My task in this project is to put some of AStUnes, a music-distribution program, into practice. It offers tools for creating playlists of songs. There are numerous ways to //add or remove music from a playlist, and there are numerous ways to combine two playlists. With the help of the classes Song.java and Playlist.java, the student is to implement AStUnes' //playlist capability. The subjects covered in the previous tasks are the main focus of this assignment, combined with Linked Lists.